Rework plugin loading into plugins.go and store actual plugins.
